SqlFilter now is an interface and defined clause generation and params binding.
This way every object that implements this interface can be used as a SQL filter for query generation and parameters binding. Even a normal `Filter` object can be also a `SqlFilter`.

On the other hand, since we cannot predict every possible use-case (JOINs, multiple different table references, etc), in `SqlFilter` there is no prescription on how a value of this type must be initialized, so we are free (and responsible) to write correct filters implementations.